Move LangService in API
