+++ test/Link.test.tsx
@@ -1,11 +1,11 @@
 import "regenerator-runtime/runtime";
 import React from "react";
-import { Link, Router, TRoute } from "../src";
+import { Link, Router } from "../src";
 import { render, fireEvent } from "@testing-library/react";
 import { Routers } from "../src/api/Routers";
 import { LangService } from "../src";
 import { TOpenRouteParams } from "../src/api/helpers";
-import { createBrowserHistory } from "history";
+import { TRoute } from "../src/components/Router";
 
 const locales = [{ key: "en" }, { key: "fr" }, { key: "de" }];
 const routesList: TRoute[] = [
@@ -14,15 +14,15 @@ const routesList: TRoute[] = [
   { path: "/bar/:id", component: null, name: "BarPage" },
 ];
 
-afterEach(() => {
-  LangService.isInit = false;
-});
-
 const mockClickHandler = jest.fn();
 const App = ({ base = "/", to }: { base: string; to: string | TOpenRouteParams }) => {
-  const history = createBrowserHistory();
+  const langService = new LangService({
+    languages: locales,
+    base,
+  });
+
   return (
-    <Router base={base} routes={routesList} history={history}>
+    <Router langService={langService} base={base} routes={routesList}>
       <Link to={to} className={"containerLink"} onClick={mockClickHandler}>
         Foo
       </Link>
@@ -46,15 +46,17 @@ describe("Link", () => {
     expect(link.getAttribute("href")).toBe("/master/foo");
   });
 
+  // FIXME
   it("should show default lang in href link", async () => {
-    LangService.init(locales, true);
+    //LangService.init(locales, true);
     const { container } = await render(<App base={"/"} to={"/foo"} />);
     const href = (container.firstChild as HTMLLinkElement).getAttribute("href");
     expect(href).toBe("/en/foo");
   });
 
+  // FIXME
   it("shouldn't show default lang in href link", async () => {
-    LangService.init(locales, false);
+    // LangService.init(locales, false);
     const { container } = await render(<App base={"/"} to={"/foo"} />);
     const href = (container.firstChild as HTMLLinkElement).getAttribute("href");
     expect(href).toBe("/foo");
