+++ test/helpers.test.ts
@@ -7,25 +7,18 @@ import {
   preventSlashes,
   removeLastCharFromString,
 } from "../src/api/helpers";
-import { Routers } from "../src";
 import { TRoute } from "../src/components/Router";
 
 // ------------------------------------------------------------ ROUTES
 
-const routes: TRoute[] = [
+const routes:TRoute[] = [
   { path: "/", component: null, name: "Home" },
   {
     path: { en: "/about", fr: "/a-propos", de: "/uber" },
-    component: { displayName: "About" } as any,
+    component: {displayName: "About"} as any,
     children: [
       { path: "/foo" },
-      {
-        path: "/bar",
-        children: [
-          { path: "/yolo/:id?", name: "YOLO" },
-          { path: "/hello", name: "Hello" },
-        ],
-      },
+      { path: "/bar", children: [{ path: "/yolo/:id?",name: "YOLO" }, { path: "/hello", name:"Hello" }] },
     ],
   },
 ];
@@ -44,8 +37,8 @@ describe("getRoutePathByRouteName", () => {
   });
 });
 
-describe("getSubRouterBase", () => {});
-describe("getSubRouterRoutes", () => {});
+describe("getSubRouterBase", () => {})
+describe("getSubRouterRoutes", () => {})
 
 // ------------------------------------------------------------ UTILS
 
@@ -112,33 +105,33 @@ describe("getUrlByPath", () => {
   });
 
   //     // FIXME besoin du LangService ici
-  //   it("should return full URL ", () => {
-  //     // LangService.init([{ key: "fr" }, { key: "en" }], true);
-  //     const routesListLang = [
-  //       {
-  //         path: "/hello",
-  //         langPath: { fr: "/salut", en: "/hello" },
-  //         children: [
-  //           {
-  //             path: "/bar-en",
-  //             langPath: { en: "/bar-en", fr: "/bar-fr" },
-  //             children: [
-  //               {
-  //                 path: "/foo-en",
-  //                 langPath: { en: "/foo-en", fr: "/foo-fr" },
-  //               },
-  //               {
-  //                 path: "/zoo",
-  //               },
-  //             ],
-  //           },
-  //           { path: "/foo" },
-  //         ],
-  //       },
-  //     ];
+//   it("should return full URL ", () => {
+//     // LangService.init([{ key: "fr" }, { key: "en" }], true);
+//     const routesListLang = [
+//       {
+//         path: "/hello",
+//         langPath: { fr: "/salut", en: "/hello" },
+//         children: [
+//           {
+//             path: "/bar-en",
+//             langPath: { en: "/bar-en", fr: "/bar-fr" },
+//             children: [
+//               {
+//                 path: "/foo-en",
+//                 langPath: { en: "/foo-en", fr: "/foo-fr" },
+//               },
+//               {
+//                 path: "/zoo",
+//               },
+//             ],
+//           },
+//           { path: "/foo" },
+//         ],
+//       },
+//     ];
 
-  //     expect(getUrlByPathPart(routesListLang, "/foo-fr")).toBe("/salut/bar-fr/foo-fr");
-  //   });
+//     expect(getUrlByPathPart(routesListLang, "/foo-fr")).toBe("/salut/bar-fr/foo-fr");
+//   });
 });
 
 describe("getUrlByRouteName", () => {
